+++ Tests/UI/NavigationUITests.swift
@@ -13,23 +13,10 @@ import XCTest
 
 @testable import Madog
 
-class NavigationUITests: MadogUIKIFTestCase {
-    private var window: UIWindow!
-    private var madog: Madog<String>!
+class NavigationUITests: MadogKIFTestCase {
     private var context: NavigationUIContext!
 
-    override func setUp() {
-        super.setUp()
-
-        window = UIWindow()
-        window.makeKeyAndVisible()
-        madog = Madog()
-        madog.resolve(resolver: KIFTestResolver())
-    }
-
     override func tearDown() {
-        window = nil
-        madog = nil
         context = nil
 
         super.tearDown()
